Refactor code to correctly use modules
