+++ index.html
@@ -68,8 +68,6 @@
 
     <footer>Copyright</footer>
 
-    <script type="module" src="./modules/date-and-time.js"></script>
-    <script type="module" src="./modules/books.js"></script>
-    <script type="module" src="./modules/navlinks-and-sections.js"></script>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
