+++ modules/books.js
@@ -1,29 +1,64 @@
-import Books, { addBookForm, small } from './../index.js'
+export const addBookForm = document.getElementById('add-books-form');
+export const bookList = document.querySelector('.book-list');
 
-const books = new Books();
-
-addBookForm.addEventListener('submit', (e) => {
-  e.preventDefault();
-
-  const title = document.querySelector('.titleField');
-  const author = document.querySelector('.authorField');
-
-  if (title.value !== '' && author.value !== '') {
-    const newBook = {
-      title: title.value,
-      author: author.value,
-    };
-
-    books.addBook(newBook);
-
-    small.innerText = 'Success! Book had been added.';
-    small.style.color = '#0bcf0b';
-    small.style.fontSize = '19px';
-  } else {
-    small.innerText = 'Please add the book\'s name & author\'s name!'
-    small.style.color = 'red';
-    small.style.fontSize = '19px';
+// Books class
+export default class Books {
+  constructor(bookLists = []) {
+    this.bookLists = bookLists;
+    this.getFromLocal();
   }
 
-  addBookForm.reset();
-});
+  removeBook(bookObject, index) {
+    const bookInfo2 = document.getElementById(index);
+    const { title, author } = bookObject;
+
+    this.bookLists = this.bookLists.filter(
+      (book) => book.title !== title && book.author !== author,
+    );
+    localStorage.setItem('booksCollection', JSON.stringify(this.bookLists));
+    bookList.removeChild(bookInfo2);
+  }
+
+  displayBook(bookObject, index) {
+    const bookInfo = document.createElement('div');
+    bookInfo.classList = 'bookInfo';
+    bookInfo.id = index;
+
+    bookInfo.innerHTML = `
+      <p class="book-details">"${bookObject.title}" by ${bookObject.author}</p>
+    `;
+
+    const removeBtn = document.createElement('button');
+    removeBtn.classList = 'remove-btn';
+    removeBtn.innerText = 'Remove';
+
+    bookInfo.appendChild(removeBtn);
+    bookList.prepend(bookInfo);
+
+    removeBtn.onclick = () => {
+      this.removeBook(bookObject, index);
+    };
+  }
+
+  addBook(bookObject) {
+    this.bookLists.push(bookObject);
+
+    localStorage.setItem('booksCollection', JSON.stringify(this.bookLists));
+
+    this.displayBook(bookObject, this.bookLists.length - 1);
+  }
+
+  // check local storage before adding a book (upon first visit to page or reload of page)
+  getFromLocal() {
+    if (localStorage.getItem('booksCollection')) {
+      this.bookLists = JSON.parse(localStorage.getItem('booksCollection'));
+
+      this.bookLists.forEach((book, index) => {
+        this.displayBook(book, index);
+      });
+    } else {
+      localStorage.setItem('booksCollection', '');
+      this.bookLists = [];
+    }
+  }
+}
