+++ modules/date-and-time.js
@@ -1,3 +1,24 @@
-import Books, { realTime } from "./../index.js"
+const displayDate = document.querySelector('.date');
 
-realTime();
+// Display the Date & Time on the DOM
+export const realTime = () => {
+  const date = new Date();
+  const dateOptions = {
+    month: 'long',
+    day: 'numeric',
+    year: 'numeric',
+  };
+
+  const timeOptions = {
+    hour: 'numeric',
+    minute: 'numeric',
+    second: 'numeric',
+    hour12: true,
+  };
+
+  const currentDate = date.toLocaleDateString('en-GB', dateOptions);
+  const currentTime = date.toLocaleTimeString('en-GB', timeOptions);
+  displayDate.innerHTML = `${currentDate} ${currentTime}`;
+
+  setTimeout(() => { realTime(); }, 1000);
+};
