+++ modules/navlinks-and-sections.js
@@ -1,16 +1,7 @@
-import { navLinks, sections } from "./../index.js"
+// Array of all navbar links
+export const navLinks = Array.from(
+  document.body.querySelectorAll('header nav ul li a'),
+);
 
-// Show or Hide Sections when a navbar link is clicked
-navLinks.forEach((navLink) => {
-  navLink.addEventListener('click', () => {
-    const idForSectionToShow = navLink.getAttribute('href');
-    sections.forEach((section) => {
-      const id = section.getAttribute('id');
-      if (`#${id}` === idForSectionToShow) {
-        section.classList.remove('hidden');
-      } else {
-        section.classList.add('hidden');
-      }
-    });
-  });
-});
+// Array of all <section> elements
+export const sections = Array.from(document.body.querySelectorAll('section'));
